# Conflicts:
#	src/main/java/com/apartment/apart/domain/notice/Notice.java
#	src/main/java/com/apartment/apart/domain/notice/NoticeController.java
#	src/main/java/com/apartment/apart/domain/report/ReportController.java
#	src/main/java/com/apartment/apart/domain/reportReply/ReportReplyController.java
#	src/main/java/com/apartment/apart/domain/reportReply/ReportReplyRepository.java
#	src/main/java/com/apartment/apart/domain/reportReply/ReportReplyService.java
#	src/main/java/com/apartment/apart/domain/schedule/ScheduleController.java
#	src/main/java/com/apartment/apart/domain/schedule/ScheduleService.java
#	src/main/java/com/apartment/apart/domain/user/SiteUser.java
#	src/main/java/com/apartment/apart/domain/user/UserRepository.java
#	src/main/java/com/apartment/apart/domain/user/UserService.java
#	src/main/java/com/apartment/apart/global/jpa/BaseEntity.java
#	src/main/resources/static/style.css
#	src/main/resources/templates/layout.html
#	src/main/resources/templates/navbar.html
#	src/main/resources/templates/report_detail.html
#	src/main/resources/templates/schedule_list.html
#	src/main/resources/templates/vote/vote_detail.html
#	src/main/resources/templates/vote/vote_list.html
#	src/main/resources/templates/vote/vote_list_list.html

+++ src/main/java/com/apartment/apart/domain/communityReply/CommunityReplyController.java
@@ -28,7 +28,7 @@ public class CommunityReplyController {
     private final UserService userService;
     @PreAuthorize("isAuthenticated()")
     @PostMapping("/create/{id}")
-    public String create(Model model, @PathVariable("id") Integer id,
+    public String create(Model model, @PathVariable("id") Long id,
                                @Valid CommunityReplyForm communityReplyForm, BindingResult bindingResult, Principal principal) {
 
         //답변 부모 질문객체를 받아온다.
